refactor(discussion/arena): hoist all loaders + signals into UseDiscussionArena
Continues the controller migration from PR #1523's hooks-and-actions
convention. Pulls every loader (`disc`, `comments`, `parent`,
`replies`, `members`), short-poll buffer (`polled_new`,
`last_seen_at`), mention plumbing (`mention_query_raw`,
`mention_query`), and derived memos (`members`, `top_priority`) up
into `UseDiscussionArena`, plus the 5-second `use_interval` polling
loop. `DiscussionArenaPage`, `CommentItem`, and `ReplyThreadView` now
destructure what they need from the cached controller instead of
each running their own `use_loader` / `use_signal` / `use_memo`
calls; `CommentItem`'s prop list shrinks from 9 to 5.

Mention autocomplete state was previously duplicated between the
arena page and the in-sheet thread view — now a single members
loader serves both composers, so a fetch from one warms the other.

Trims verbose history-narrating comments throughout the file in the
same pass; section dividers in RSX and field-by-field doc comments
in the controller struct were doing more harm than good. Surviving
comments cover only non-obvious WHY (default-on-None loader trick,
SSR/hydration hook order, polled snapshot patches).
